requirements.txt
Authlib==1.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cycler==0.11.0
dnspython==2.2.1
email-validator==1.1.3
filelock==3.6.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.33.3
greenlet==1.1.2
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
MarkupSafe==2.1.1
matplotlib==3.5.2
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pycparser==2.21
pynndescent==0.5.6
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.35
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
typing-extensions==4.2.0
umap-learn==0.5.3
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
